requirements/production.txt
Django==1.10.0
djangorestframework >= 3.5,<3.6
django-filter==1.0.4
mysqlclient
django-mptt
Pillow
gunicorn
requests
pycrypto
qrcode==5.1